pyproject.toml
[build-system]
requires = ["setuptools>=69.0.0",]
build-backend = "setuptools.build_meta"

[project]
authors = [
	{name = "E God (evnchn)"},
	{name = "Happypig375", email = "hadrianwttang@outlook.com"},
	{name = "polyipseity", email = "polyipseity@gmail.com"},
]
classifiers = [
	"Development Status :: 5 - Production/Stable",
	"Environment :: Console",
	"Framework :: AnyIO",
	"Framework :: AsyncIO",
	"Framework :: aiohttp",
	"Intended Audience :: Information Technology",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Topic :: Internet",
	"Topic :: Internet :: WWW/HTTP",
	"Topic :: Internet :: WWW/HTTP :: Indexing/Search",
	"Typing :: Typed",
]
dependencies = [
	"aiohttp>=3.9.0", # async request
	"anyio>=4.3.0", # async file
	"asyncstdlib>=3.12.0", # async utilities
	"beautifulsoup4>=4.12.0", # parse HTML
	"nicegui>=1.4.0", # web interface
	"nltk>=3.8.0", # natural language processing, in particular for splitting words
	"tortoise-orm>=0.20.0", # async SQLite database, ORM-style
	"tqdm>=4.66.0", # progress bar
	"yarl>=1.9.0", # URL utilities
]
description = "HKUST COMP 4321 sources dropped from heaven by E God himself!"
dynamic = ["version",]
keywords = []
license = {file = "LICENSE.md"}
name = "egod-search" # synchronize with `__init__.py`
readme = "README.md"
requires-python = ">=3.11.0"

[project.optional-dependencies]
all = ["egod_search[dev]"]
dev = [
	"pyinstaller>=6.5.0", # build native executables
	"unittest-parallel>=1.6.0", # `unittest` with coverage and parallelism
]

[project.scripts]
egod-search-build = "egod_search._build:main"
egod-search-crawl = "egod_search.crawl.__main__:main"
egod-search-test = "egod_search._test:main"

[project.urls]
repository = "https://github.com/polyipseity/comp-4321.git"

[tool.setuptools.dynamic]
version = {attr = "egod_search.VERSION"}
